compute in floats

+++ src/fuzzy_search.h
@@ -6,7 +6,7 @@
 
 #include "error.h"
 
-extern zsql_error *fuzzy_search(double *score, const int32_t *haystack,
+extern zsql_error *fuzzy_search(float *score, const int32_t *haystack,
                                 size_t haystack_length, const int32_t *needle,
                                 size_t needle_length);
